Split run_bench.py into runner.py, fixtures.py, and thin entry point
run_bench.py was approaching 530 lines with three distinct concerns mixed
together. Extract the execution engine to runner.py (with a clean
non-argparse API), the fixture blob to fixtures.py, and leave run_bench.py
as a pure CLI dispatcher. Add tests/bench/test_runner.py with 16 tests
covering _is_oom, _make_driver, _make_prompt, and _case_id.
